clean up a bit
